refactor: /stats extract components
